Clean up religious usage of size_t; use int instead. (#22)
* begin string cleanups and remove many unneeded size_t uses

* Clean up religious usage of size_t; use int instead.

* fix mkdir breaking on windows

* makefile bugfix for windows

+++ z/core/binaryStream.cpp
@@ -20,10 +20,10 @@ namespace z
 			data.add(ch);
 		}
 
-		void binaryStream::put(uint8_t* str, size_t count, encoding format)
+		void binaryStream::put(uint8_t* str, int count, encoding format)
 		{
 			if (!str) return;
-			size_t datact;
+			int datact;
 
 			switch (format)
 			{
@@ -39,7 +39,7 @@ namespace z
 				datact = count;
 			}
 
-			for (size_t i=0; i<datact; i++)
+			for (int i=0; i<datact; i++)
 			{
 				data.add(str[i]);
 			}
@@ -47,7 +47,7 @@ namespace z
 
 		uint8_t binaryStream::get()
 		{
-			if (streamIndex < data.length())
+			if (streamIndex < (size_t)data.length())
 				return data[streamIndex++];
 			else
 				return 0;
@@ -55,9 +55,9 @@ namespace z
 
 		uint32_t binaryStream::getChar()
 		{
-			if (streamIndex >= data.length()) return 0;
+			if (streamIndex >= (size_t)data.length()) return 0;
 
-			size_t datact;
+			int datact;
 
 			uint8_t res8;
 			uint16_t res16;
@@ -81,13 +81,13 @@ namespace z
 				str = (uint8_t*)&res8;
 			}
 
-			if (streamIndex+datact > data.length())
+			if (streamIndex+datact > (size_t)data.length())
 			{
 				streamIndex = data.length();
 				return 0;
 			}
 
-			for (size_t i=0; i<datact; i++)
+			for (int i=0; i<datact; i++)
 				str[i] = data[streamIndex++];
 
 			switch (format())
@@ -103,7 +103,7 @@ namespace z
 
 		bool binaryStream::empty()
 		{
-			return (streamIndex >= data.length());
+			return (streamIndex >= (size_t)data.length());
 		}
 
 		bool binaryStream::good()
@@ -128,7 +128,7 @@ namespace z
 
 		void binaryStream::seek(size_t index)
 		{
-			if (index > data.length())
+			if (index > (size_t)data.length())
 				streamIndex = data.length();
 			else
 				streamIndex = index;
@@ -151,13 +151,13 @@ namespace z
 
 			if (this->empty()) return streamFormat = ascii;
 
-			size_t read_max = (32 > data.length()) ? data.length() : 32;
+			int read_max = (32 > data.length()) ? data.length() : 32;
 
-			size_t contig_nulls = 0;
+			int contig_nulls = 0;
 			bool found_nulls = false;
 			bool can_utf8 = true;
 
-			for (size_t i=0; i<read_max; i++)
+			for (int i=0; i<read_max; i++)
 			{
 				//ascii and utf8 won't have null chars
 				if (found_nulls || !data[i])
