+++ z/util/regex/rules/space.cpp
@@ -6,7 +6,7 @@ namespace z
 	{
 		namespace rgx
 		{
-			space::space(bool negate, size_t min, size_t max, bool greedy):
+			space::space(bool negate, int min, int max, bool greedy):
 				rule(min,max,greedy), negate(negate) {}
 
 			bool space::match(uint32_t current) const
