+++ z/util/regex/rules/word.cpp
@@ -6,7 +6,7 @@ namespace z
 	{
 		namespace rgx
 		{
-			word::word(bool negate, size_t min, size_t max, bool greedy):
+			word::word(bool negate, int min, int max, bool greedy):
 				rule(min,max,greedy), negate(negate) {}
 
 			bool word::match(uint32_t current) const
